Reduce MessageList copypaste
