+++ src/components/MessageList.js
@@ -1,18 +1,11 @@
 /* eslint-disable react/prop-types */
 /* eslint-disable react/destructuring-assignment */
 import React from 'react'
-import {
-	useParams
-} from "react-router-dom";
-import { connect } from 'react-redux';
 import useForm from 'react-hook-form';
 import styled from '@emotion/styled';
 import Message from './MessageForm';
 import Input from './MessageInput';
 import EmojiList from './EmojiList';
-import { getTime } from '../actions/time';
-import { saveMessage } from '../actions/localDb';
-import { getProfile } from '../actions/index';
 
 const Container = styled.div`
 	display: flex;
@@ -31,77 +24,81 @@ const MessageContainer = styled.div`
 	background-color: rgba(0, 0, 0, 0.04);
 `;
 
-function MessageList(props) {
+export function EmptyMessageList(props, chatId, onSubmit, variableName, messages, setMessages) {
 	const {
 		inputValue,
 		setInputValue,
-		localMessages,
-		setLocalMessages,
+		yourName,
+		setYourName,
+		setChats,
+		userName,
+		isRecording,
+		setIsRecording,
+		setUserName,
+		emojiMode,
+		setEmojiMode,
+	} = props.state;
+	const { register, handleSubmit } = useForm();
+	const handleYourMessage = () => setYourName(userName);
+	const handleCompanionMessage = () => setYourName('Companion');
+	const handleChange = (event) => setInputValue(event.target.value);
+	const handleYourNameChange = (event) => setUserName(event.target.value);
+	const changeEmojiMode = () => setEmojiMode(!emojiMode);
+	return (
+		<Container>
+			<MessageContainer id='messageContainer'/>
+			<Input
+				onSubmit={handleSubmit(onSubmit)}
+				onChange={handleChange}
+				setInputValue={setInputValue}
+				value={inputValue}
+				youOnTyping={handleYourMessage}
+				compOnTyping={handleCompanionMessage}
+				changeYourName={handleYourNameChange}
+				setMessages={setMessages}
+				yourName={yourName}
+				setChats={setChats}
+				chatId={chatId}
+				messages={messages}
+				setIsRecording={setIsRecording}
+				isRecording={isRecording}
+				refer={register}
+				variableName={variableName}
+				changeEmojiMode={changeEmojiMode}
+			/>
+			<EmojiList
+				emojiMode={emojiMode}
+				inputValue={inputValue}
+				setInputValue={setInputValue}
+			/>
+		</Container>
+	);
+}
+
+export function MessageList(props, chatId, onSubmit, variableName, userName, messages, setMessages) {
+	const {
+		inputValue,
+		setInputValue,
 		yourName,
 		setYourName,
 		setChats,
 		setMessagesEnd,
 		isRecording,
 		setIsRecording,
+		setUserName,
 		emojiMode,
 		setEmojiMode,
 	} = props.state;
-	const {
-		userName,
-	} = props;
-	// eslint-disable-next-line react/prop-types
-	const { chatId } = useParams();
 	const { register, handleSubmit } = useForm();
-	const onSubmit = (values) => {
-		if (inputValue !== '') {
-			const curTime = getTime();
-			const message = {
-				name: yourName,
-				chat_id: Number(chatId),
-				added_at: curTime,
-				content: inputValue,
-			}
-			saveMessage(message, true, setLocalMessages, setChats, localMessages);
-			setInputValue('');
-		}
-	};
 	const handleYourMessage = () => setYourName(userName);
 	const handleCompanionMessage = () => setYourName('Companion');
 	const handleChange = (event) => setInputValue(event.target.value);
+	const handleYourNameChange = (event) => setUserName(event.target.value);
 	const changeEmojiMode = () => setEmojiMode(!emojiMode);
-	if (localMessages === null) {
-		return (
-			<Container>
-				<MessageContainer id='messageContainer'/>
-				<Input
-					onSubmit={handleSubmit(onSubmit)}
-					onChange={handleChange}
-					setInputValue={setInputValue}
-					value={inputValue}
-					youOnTyping={handleYourMessage}
-					compOnTyping={handleCompanionMessage}
-					setMessages={setLocalMessages}
-					yourName={yourName}
-					setChats={setChats}
-					chatId={chatId}
-					messages={localMessages}
-					setIsRecording={setIsRecording}
-					isRecording={isRecording}
-					refer={register}
-					changeEmojiMode={changeEmojiMode}
-				/>
-				<EmojiList
-					emojiMode={emojiMode}
-					inputValue={inputValue}
-					setInputValue={setInputValue}
-				/>
-			</Container>
-		);
-	}
 	return (
 		<Container>
 			<MessageContainer id='messageContainer'>
-				{localMessages
+				{messages
 					.filter((message) => message.chat_id === Number(chatId))
 					.map((message) => (
 						<Message
@@ -126,14 +123,16 @@ function MessageList(props) {
 				value={inputValue}
 				youOnTyping={handleYourMessage}
 				compOnTyping={handleCompanionMessage}
-				setMessages={setLocalMessages}
+				changeYourName={handleYourNameChange}
+				setMessages={setMessages}
 				yourName={yourName}
 				setChats={setChats}
 				chatId={chatId}
-				messages={localMessages}
+				messages={messages}
 				setIsRecording={setIsRecording}
 				isRecording={isRecording}
 				refer={register}
+				variableName={variableName}
 				changeEmojiMode={changeEmojiMode}
 			/>
 			<EmojiList
@@ -144,12 +143,3 @@ function MessageList(props) {
 		</Container>
 	);
 }
-
-const mapStateToProps = (state) => ({
-	userName: state.profile.userName,
-})
-
-export default connect(
-	mapStateToProps,
-	{ getProfile },
-)(MessageList)
